+++ src/Config.h
@@ -91,6 +91,7 @@ public:
     double      buildingPercentile              = -9999.; // Handled by schema
     double      minHeight                       = 2.;
     bool        reconstructFailed               = false;
+    std::string crsInfo;                                  // CRS information of building footprints
     // Height from attributes
     std::string buildingHeightAttribute;
     std::string floorAttribute;
@@ -124,7 +125,6 @@ public:
     std::string        logName                  = std::string("log");
     std::ostringstream log;
     std::ostringstream logSummary;
-    std::vector<int>   failedBuildings;
 
     //-- Experimental
     bool       clip                             = false;