+++ src/io.cpp
@@ -30,13 +30,15 @@
 #include "Config.h"
 #include "TopoFeature.h"
 #include "Boundary.h"
+#include "Building.h"
 
 #include <CGAL/Polygon_mesh_processing/compute_normal.h>
 #include <CGAL/Polygon_mesh_processing/connected_components.h>
 #include <CGAL/Polygon_mesh_processing/transform.h>
-#include <CGAL/Polygon_mesh_processing/repair_polygon_soup.h>
 #include <CGAL/Polygon_mesh_processing/IO/polygon_mesh_io.h>
 #include <CGAL/IO/read_las_points.h>
+
+#include <ogrsf_frmts.h>
 
 #include <boost/algorithm/string.hpp>
 
@@ -114,6 +116,150 @@ void IO::read_geojson_polygons(std::string& file, JsonVectorPtr& jsonPolygons) {
     } catch (std::exception& e) {
         throw std::runtime_error(std::string("Error parsing JSON file '" + file + "'. Details: " + e.what()));
     }
+}
+
+void IO::read_polygons(std::string& file, PolyVecPtr& polygons, std::string* crsInfo) {
+    // Prepare input map
+    GDALAllRegister();
+    GDALDataset *inputMapDataset = (GDALDataset*) GDALOpenEx(file.c_str(), GDAL_OF_READONLY, NULL, NULL, NULL);
+    if (inputMapDataset == NULL) {
+        throw std::runtime_error("Error: Could not open input polygon");
+    }
+    std::cout << "GDAL: Reading polygon file: " << file << " type: " << inputMapDataset->GetDriverName() << std::endl;
+
+    CPLStringList metadataDomains(inputMapDataset->GetMetadataDomainList());
+    for (int currentDomain = 0; currentDomain < metadataDomains.Count(); ++currentDomain) {
+        //    std::cout << metadataDomains[currentDomain] << std::endl;
+    }
+
+    // Load map polygons
+    //todo maybe a warning when there are multiple layers?
+    for (auto &&inputLayer: inputMapDataset->GetLayers()) {
+        inputLayer->ResetReading();
+
+        // Try to extract CRS from this layer
+        OGRSpatialReference *spatialRef = inputLayer->GetSpatialRef();
+        if (spatialRef != NULL) {
+            std::string crsAuthority, crsCode;
+            //      char *srs = (char *)CPLMalloc(10000*sizeof(char));
+            //      spatialRef->exportToPrettyWkt(&srs);
+            //      std::cout << srs << std::endl;
+            //      CPLFree(srs);
+            const char *authority = spatialRef->GetAuthorityName(NULL);
+            const char *code = spatialRef->GetAuthorityCode(NULL);
+            if (authority != NULL) crsAuthority = std::string(authority);
+            if (code != NULL) crsCode = std::string(code);
+            if (crsInfo) *crsInfo += crsAuthority + "::" + crsCode;
+            std::cout << "    CRS information: " << authority << " code: " << code << std::endl;
+        }
+
+        OGRFeature *inputFeature;
+        while ((inputFeature = inputLayer->GetNextFeature()) != NULL) {
+            if (!inputFeature->GetGeometryRef()) continue;
+//            std::string cityjson_class;
+            std::unordered_map<std::string, std::string> attributes;
+
+            for (int currentField = 0; currentField < inputFeature->GetFieldCount(); ++currentField) {
+//                if (strcmp(inputFeature->GetFieldDefnRef(currentField)->GetNameRef(), class_attribute.c_str()) == 0) {
+//                    cityjson_class = inputFeature->GetFieldAsString(class_attribute.c_str());
+//                    continue;
+//                }
+                if (inputFeature->IsFieldNull(currentField)) continue;
+                switch (inputFeature->GetFieldDefnRef(currentField)->GetType()) {
+                    case OFTReal:
+                        attributes[inputFeature->GetFieldDefnRef(currentField)->GetNameRef()]
+                                = std::to_string(inputFeature->GetFieldAsDouble(currentField));
+                        break;
+                    case OFTString:
+                        attributes[inputFeature->GetFieldDefnRef(currentField)->GetNameRef()]
+                                = inputFeature->GetFieldAsString(currentField);
+                        break;
+                    case OFTInteger:
+                        attributes[inputFeature->GetFieldDefnRef(currentField)->GetNameRef()]
+                                = std::to_string(inputFeature->GetFieldAsInteger(currentField));
+                        break;
+                    case OFTInteger64:
+                        attributes[inputFeature->GetFieldDefnRef(currentField)->GetNameRef()]
+                                = std::to_string(inputFeature->GetFieldAsInteger64(currentField));
+                        break;
+                    default:
+//                        std::cout << "Unsupported field type: " << inputFeature->GetFieldDefnRef(currentField)->GetType() << std::endl;
+                        break;
+                }
+            }
+
+            if (wkbFlatten(inputFeature->GetGeometryRef()->getGeometryType()) == wkbPolygon ||
+                wkbFlatten(inputFeature->GetGeometryRef()->getGeometryType()) == wkbTriangle) {
+                OGRPolygon *inputPolygon = inputFeature->GetGeometryRef()->toPolygon();
+                polygons.emplace_back(std::make_unique<Polygon_with_attr>());
+//                polygons.back()->id = std::to_string(inputFeature->GetFID());
+//                polygons.back().semantic_class = cityjson_class;
+                polygons.back()->attributes = attributes;
+                polygons.back()->polygon.rings().emplace_back();
+                for (int currentVertex = 0;
+                     currentVertex < inputPolygon->getExteriorRing()->getNumPoints();
+                     ++currentVertex) {
+                    polygons.back()->polygon.rings().back()
+                            .push_back(Point_2(inputPolygon->getExteriorRing()->getX(currentVertex)
+                                               - Config::get().pointOfInterest.x(),
+                                               inputPolygon->getExteriorRing()->getY(currentVertex)
+                                               - Config::get().pointOfInterest.y()));
+                }
+                for (int currentInnerRing = 0;
+                     currentInnerRing < inputPolygon->getNumInteriorRings();
+                     ++currentInnerRing) {
+                    polygons.back()->polygon.rings().emplace_back();
+                    for (int currentVertex = 0;
+                         currentVertex < inputPolygon->getInteriorRing(currentInnerRing)->getNumPoints();
+                         ++currentVertex) {
+                        polygons.back()->polygon.rings().back()
+                                .push_back(Point_2(inputPolygon->getInteriorRing(currentInnerRing)->getX(currentVertex)
+                                                   - Config::get().pointOfInterest.x(),
+                                                   inputPolygon->getInteriorRing(currentInnerRing)->getY(currentVertex)
+                                                   - Config::get().pointOfInterest.y()));
+                    }
+                }
+            } else if (wkbFlatten(inputFeature->GetGeometryRef()->getGeometryType()) == wkbMultiPolygon) {
+                OGRMultiPolygon *inputMultipolygon = inputFeature->GetGeometryRef()->toMultiPolygon();
+                for (int currentPolygon = 0;
+                     currentPolygon < inputMultipolygon->getNumGeometries();
+                     ++currentPolygon) {
+                    OGRPolygon *inputPolygon = inputMultipolygon->getGeometryRef(currentPolygon);
+                    polygons.push_back(std::make_unique<Polygon_with_attr>());
+//                    polygons.back()->id = std::to_string(inputFeature->GetFID()) + "-" + std::to_string(currentPolygon);
+//                    polygons.back().semantic_class = cityjson_class;
+                    polygons.back()->attributes = attributes;
+                    polygons.back()->polygon.rings().emplace_back();
+                    for (int currentVertex = 0;
+                         currentVertex < inputPolygon->getExteriorRing()->getNumPoints();
+                         ++currentVertex) {
+                        polygons.back()->polygon.rings().back()
+                                .push_back(Point_2(inputPolygon->getExteriorRing()->getX(currentVertex)
+                                                   - Config::get().pointOfInterest.x(),
+                                                   inputPolygon->getExteriorRing()->getY(currentVertex)
+                                                   - Config::get().pointOfInterest.y()));
+                    }
+                    for (int currentInnerRing = 0;
+                         currentInnerRing < inputPolygon->getNumInteriorRings();
+                         ++currentInnerRing) {
+                        polygons.back()->polygon.rings().emplace_back();
+                        for (int currentVertex = 0;
+                             currentVertex < inputPolygon->getInteriorRing(currentInnerRing)->getNumPoints();
+                             ++currentVertex) {
+                            polygons.back()->polygon.rings().back()
+                                    .push_back(Point_2(inputPolygon->getInteriorRing(currentInnerRing)->getX(currentVertex)
+                                                       - Config::get().pointOfInterest.x(),
+                                                       inputPolygon->getInteriorRing(currentInnerRing)->getY(currentVertex)
+                                                       - Config::get().pointOfInterest.y()));
+                        }
+                    }
+                }
+            } else {
+                //        std::cout << "Ignoring non-areal type..." << std::endl;
+            }
+        }
+    }
+    GDALClose(inputMapDataset);
 }
 
 void IO::read_other_geometries(std::string& file, std::vector<Mesh>& meshes) {
@@ -387,32 +533,58 @@ void IO::output_log() {
     if (!Config::get().outputLog) return;
     fs::current_path(Config::get().outputDir);
 
-    //-- Output log file
     Config::get().log <<"\n// ------------------------------------------------------------------------------------------------ //" << std::endl;
     std::cout << "\nCreating log file '" << Config::get().logName << "'" << std::endl;
     std::ofstream of;
     of.open(Config::get().logName);
     of << Config::get().logSummary.str() << Config::get().log.str();
     of.close();
+}
+
+void IO::output_log(const BuildingsPtr failedBuildings) {
+    if (!Config::get().outputLog) return;
+    fs::current_path(Config::get().outputDir);
+
+    //-- Output log file
+    IO::output_log();
 
     //-- Output failed reconstructions
-    if (!Config::get().failedBuildings.empty()) {
+    if (!failedBuildings.empty()) {
         std::cout << "Outputting failed building reconstructions to 'failedReconstructions.geojson'"
                   << std::endl;
-        //- Parse again the buildings polygon
-        std::ifstream ifs(Config::get().workDir.append(Config::get().gisdata).string());
-        nlohmann::json j = nlohmann::json::parse(ifs);
-        //- Extract failed reconstructions
-        nlohmann::json b;
-        for (int i: Config::get().failedBuildings) {
-            b["features"].push_back(j["features"][i]);
+        nlohmann::json j;
+        //-- Header
+        if (!Config::get().crsInfo.empty())
+            j["crs"]["properties"]["name"] = "urn:ogc:def:crs:" + Config::get().crsInfo;
+        else
+            j["crs"]["properties"]["name"] = "";
+        j["crs"]["type"] = "name";
+        j["name"] = "failedBuildings";
+        j["type"] = "featureCollection";
+        //-- Features
+        for (auto f : failedBuildings) {
+            nlohmann::json b;
+            b["type"] = "Feature";
+            b["geometry"]["type"] = "Polygon";
+            b["properties"]["gid"] = f->get_id();
+            bool outerRing = true;
+            for (auto ring : f->get_poly().rings()) {
+                if (outerRing) outerRing = false;
+                else ring.reverse_orientation();
+                nlohmann::json c;
+                for (int i = 0; i < ring.size() + 1; ++i) {
+                    c.emplace_back();
+                    c.back().push_back(ring[i % ring.size()].x() + Config::get().pointOfInterest.x());
+                    c.back().push_back(ring[i % ring.size()].y() + Config::get().pointOfInterest.y());
+                }
+                b["geometry"]["coordinates"].push_back(c);
+            }
+            j["features"].push_back(b);
         }
-        b["crs"] = j["crs"];
-        b["name"] = "failedBuildings";
-        b["type"] = j["type"];
         //- Write to file
+        std::ofstream of;
         of.open("failedReconstructions.geojson");
-        of << b.dump();
+        of << j.dump();
         of.close();
     }
 }