+++ src/main.cpp
@@ -133,7 +133,7 @@ int main(int argc, char** argv) {
         map3d.output();
 
         //-- Output log
-        IO::output_log();
+        IO::output_log(map3d.get_failed_buildings());
 
         auto endTime = std::chrono::steady_clock::now();
         auto diffTime = endTime - startTime;